organize fallbacks, homepage scale
