--- frontend/src/data/fallbackNFT.ts
@@ -1,16 +1,0 @@
-import { NFT } from "../types/NFT"
-import { RarityLevel } from "../types/Rarity"
-
-export const fallbackNFT: NFT = {
-    id: 9999,
-    image: "QmSmZ7GBc3TssiZnZRazxCBSeenL2iSYmfQtNELGLCfg9d",
-    description: "This is a fallback NFT image.",
-    attributes: [
-        {
-            trait_type: "Rarity",
-            value: 1,
-        },
-    ],
-}
-
-export const fallbackNFTRarityLevel = RarityLevel.Legendary
