+++ frontend/src/data/mock.ts
@@ -1,0 +1,24 @@
+import { MetadataMockProperty } from "../components/ui/MetadataMock"
+
+export const mockMetaBannerData: MetadataMockProperty[] = [
+    {
+        name: "image",
+        val: "QmQiHyuPLdC49vKDt1rJ1iyFC4SpFFmF7yoDWEXWPfRF7Z",
+        comment: "IPFS cid of NFT",
+    },
+    {
+        name: "rarity",
+        val: "681472",
+        comment: "NFT rarity with 6 digits of precision",
+    },
+    {
+        name: "season",
+        val: "Summer 2024",
+        comment: "Season during which the NFT was created",
+    },
+    {
+        name: "description",
+        val: "Generate a minimalistic portrait of a fictional character. Use a solid color background. The main features of this character are: cyborg, human, android",
+        comment: "The prompt used for generation",
+    },
+]
